slouch bot is not interpreting any messages
I looked at the source for receiving messages, and it looks for it's
name or id or "mention" (@id), so I printed those out to make sure I was
addressing it correctly. I was, however, even with logging set to DEBUG
level, it still did not correctly identify (and log) messages with it's
name in them.